@@ -9,16 +9,20 @@ import { TimepickerModule } from 'ngx-bootstrap/timepicker';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ClientsPageComponent } from './clients-page/clients-page.component';
-import { TasksComponent } from './clients-page/tasks/tasks.component';
 import { LoginPageComponent } from './login-page/login-page.component';
 import { RegisterPageComponent } from './register-page/register-page.component';
 import { TokenInterceptor } from './shared/classes/token.interceptor';
 import { AuthLayoutComponent } from './shared/layouts/auth-layout/auth-layout.component';
-import { ClientStatisticComponent } from './shared/layouts/clients-layout/clients-statistic.component';
 import { SiteLayoutComponent } from './shared/layouts/site-layout/site-layout.component';
-import { SharedModule } from './shared/module';
 import { AppStoreModule } from './store/app-store.module';
+
+import { ClientsPageComponent } from './clients-page/clients-page.component';
+import { TasksComponent } from './clients-page/tasks/tasks.component';
+import { ReportsComponent } from './reports/reports.component';
+import { ClientStatisticComponent } from './shared/layouts/clients-layout/clients-statistic.component';
+import { SharedModule } from './shared/module/shared-module/shared-module.module';
+import { ClientReportsComponent } from './reports/client-reports/client-reports.component';
+
 
 @NgModule({
   declarations: [
@@ -29,6 +33,8 @@ import { AppStoreModule } from './store/app-store.module';
     RegisterPageComponent,
     ClientsPageComponent,
     TasksComponent,
+    ReportsComponent,
+    ClientReportsComponent,
     ClientStatisticComponent,
   ],
   imports: [
