@@ -3,6 +3,7 @@ const jwt = require("jsonwebtoken");
 const User = require("../models/User");
 const keys = require("../config/keys");
 const errorHandler = require("../utils/errorHandler");
+const path = require('path');
 
 module.exports.getUser = async function(req, res) {
   const candidate = await User.findOne({ _id: req.params.id });
@@ -52,6 +53,7 @@ module.exports.login = async function(req, res) {
     });
   }
 };
+
 module.exports.register = async function(req, res) {
   // email password
   const candidate = await User.findOne({ email: req.body.email });
@@ -68,7 +70,7 @@ module.exports.register = async function(req, res) {
     const user = new User({
       email: req.body.email,
       password: bcrypt.hashSync(password, salt),
-      imageSrc: req.file ? req.file.path : "",
+      imageSrc: req.file ? req.file.path : "uploads\\default.png",
       role: req.body.role,
       nickName: req.body.nickName
     });
