@@ -42,4 +42,4 @@ const taskSchema = new Schema({
 	}
 });
 
-module.exports = mongoose.model("tasks", taskSchema);
+module.exports = mongoose.model("archivedTasks", taskSchema, "archivedTasks");
